requirements.txt
fastembed 
qdrant-client
datasets
transformers
python-dotenv
PyPDF2
langchain-core
cohereaiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
boto3==1.34.119
botocore==1.34.119
certifi==2024.6.2
charset-normalizer==3.3.2
cohere==5.5.4
coloredlogs==15.0.1
datasets==2.18.0
dill==0.3.8
fastavro==1.9.4
fastembed==0.2.7
filelock==3.14.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.2.0
grpcio==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.20.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain-core==0.2.4
langsmith==0.1.71
loguru==0.7.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.18.0
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pillow==10.3.0
portalocker==2.8.2
protobuf==5.27.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.7.3
pydantic_core==2.18.4
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.9.1
regex==2024.5.15
requests==2.32.3
s3transfer==0.10.1
safetensors==0.4.3
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sympy==1.12.1
tenacity==8.3.0
tokenizers==0.15.2
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.39.3
types-requests==2.32.0.20240602
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
wheel==0.43.0
xxhash==3.4.1
yarl==1.9.4
psutil
